refactor(amount): separate amount to mod
